Fix warnings, now FILE->_cnt is a ssize_t.

+++ tests/libc/ansi/stdio/filbuf.c
@@ -12,8 +12,8 @@ main(void)
   {
     if (f->_cnt > last_count)
     {
-      printf("Count: %5d  Fill: %5lu  Bufsiz: %5lu  SoFar: %5d\n",
-	     f->_cnt, (unsigned long) f->_fillsize,
+      printf("Count: %5lu  Fill: %5lu  Bufsiz: %5lu  SoFar: %5d\n",
+	     (unsigned long) f->_cnt, (unsigned long) f->_fillsize,
 	     (unsigned long) f->_bufsiz, so_far);
     }
     last_count = f->_cnt;
